[REM] mrp_workorder: remove dead code
The variable `main_employee_connected` was previously used
inside the `_set_default_time_log` method, but it is no longer used
since it was removed in this [PR](https://github.com/odoo/enterprise/pull/84790).

As a result, the variable is redundant and has no functional impact.

This commit removes the dead code to keep the codebase clean and maintainable.

closes odoo/enterprise#93510

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>