[IMP] *: gantt: multi selection with ctrl
We make possible to use Control to select several (blocks of) cells when
the multi create mode is enabled in the gantt view. The rules are
basically the same as the one used for selecting files on various
operating systems.  This allows to quickly add/delete records in several
regions at once.

We adapt the work_entries_gantt view to those changes and the changes
brought by the commit [REF] *: simplify replacement of work entries in
https://github.com/odoo/odoo/pull/223409.
Along the way, we have refactor some parts of that view. Most notably,
we no longer fetch the serFavoritesWorkEntriescts in onWillRender.

Task ID: 5003199

closes odoo/enterprise#92641

Related: odoo/odoo#223409
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>