[REF] mrp_mps: refactor mps suggestion to a wizard
This commit refactors the MPS suggestion feature by making it a
wizard.

This also helps fix the problem with the selection field
`suggestion_period` in the MPS model as storing this selection field
as False in the database leads to crashes when MPS is uninstalled.

related PR: https://github.com/odoo/enterprise/pull/90553

closes odoo/enterprise#93524

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>