[FIX] documents: prevent pypdf warning in tests
Following the activation by default of account document synchronization, a lot
of tests now create indirectly documents. In test mode,
IrActionsReport._pre_render_qweb_pdf renders pdf as html but with the pdf mime
type leading to a warning when this fake pdf is provided to
documents_document._get_is_multipage method. In odoo/enterprise#90662, we have
prevented _get_is_multipage to emit a warning by turning it in an info log when
in testing mode, and we detect that the document content is actually a html
content. Unfortunately, at that time, pypdf was not emitting a warning so we
have done that check when an error happens. To solve the problem we check the
content before sending it to pypdf to avoid the pypdf warning.

Task-5052041

closes odoo/enterprise#93537

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>