[FIX] l10n_br_edi_pos, pos_*: prevent missing recursive infinite loop
*: pos_settle_due, pos_urban_piper

The `pos.order` model was missing from pohibitedAutoLoadedModels which
was causing infinite loop when validating an order.

This commit add this models to the pohibitedAutoLoadedModels variable.

Now when downloading an order from the PoS the `read_pos_orders`
method should be used, it will returns all related order data.

closes odoo/enterprise#93726

X-original-commit: 6319e03f9987ac3c869d3535c9dd2e77d4506404
Related: odoo/odoo#225172
Signed-off-by: David Monnom (moda) <moda@odoo.com>