[FIX] account_reports: remove related field report_country_id
this field is no longer needed since the field country_id contains
the necessary information.

Also, makes the field country_id readonly=False, it was
originally the case and got changed when switched to a computed
field

Finally, change the recently added field auto_generate to a computed
field since we want the default value to be True if it is of
category "Tax Return".

closes odoo/enterprise#93535

Related: odoo/upgrade#8327
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>