[FIX] account_online_synchronization: prevent conflict with lock dates
If you have a company with a lock date set, and you connect a bank that has
transactions dated to the day of the lock date, these transactions will be
fetched, which do not make sense because nothing should be created in a period
covered by a lock date.

As a result, these transactions would be created at a wrong date (if the current
month is the first after the lock date or if the sequence has a monthly reset,
then they would be appended to the current month, else if the sequence reset
annually, then they would be created at the current date).

Additionally, the potential opening balance would be created at a wrong date
too, since it would try to create it one day prior to the oldest transaction.

The date which the opening balance is created would not be the same as the
transactions above, which adds a layer to the mess created.

To prevent this, at initialization, we set the last sync date one day after the
lock date, not the same day.

As for the opening balance, we do not try to set it one day prior to the oldest
transaction, but the same day. The `internal_index` computed will ensure it is
displayed as the first transaction of that journal.

Finally, the test related to statement creation were adapted to this new
behavior.

opw-4890538

closes odoo/enterprise#98307

X-original-commit: 07a485160d80df4c76f82c3fbeb3d379f5e8186f
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>