[FIX] timesheet_grid: hide timesheets tab in task form view when create and edit from timer header
Before this commit, when the user starts a timer from the timer header
inside a timesheet view and create and edit a task from `task_id` field
inside the timer header, if the user creates a timesheet in the form
view of task, the modal containing the form view immediately closes.
The reason is because the adding of timesheet of the new task conflicts
with the behavior of `create and edit` of task from timesheet record
(inside timer header in that case).

This commit hides the timesheets tab inside the form view of task when
the user creates and edits a task from timer header.

task-4284216

Part-of: odoo/enterprise#81686
Related: odoo/odoo#202313
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>