[FIX] timesheet_grid: stop timer in row when timer stopped in timer header
Before this commit, when the user starts a timer in the grid view and go
to previous period. If the row containing the timer is also displayed in
the previous period and the user stops the timer. The row still show the
timer is running even if the timer is stopped.

This commit correctly notifies the row the timer is stopped when the
user clicks on stop button in the timer header.

task-4284216

closes odoo/enterprise#81686

Related: odoo/odoo#202313
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>