[FIX] iot: handle blackbox serial port change
Before this commit, if a blackbox was unplugged and re-plugged, and it
was assigned a different serial port by the Raspberry Pi, it would show
up as a new device in the database meaning the existing configuration
wouldn't work.

After this commit, we handle the blackbox as a special case, and if the
name of the device matches exactly with our existing blackbox, we update
its identifier instead of creating a new device.

This does require a new device specific check in the controller which is
quite ugly. Another approach would have been to make the identifier of
the blackbox equal its FDM ID instead of the serial port, but this was
not done for the following reasons:
- Changing the identifier format in stable would cause all existing
  clients' blackboxes to become unconfigured once their IoT box
  restarts.
- Making the identifier different to the serial port would require a
  hack in the blackbox driver to change its own identifier and update
  the devices dictionary, since the serial interface assumes all devices
  use the port as their identifier.

task-5055027

closes odoo/enterprise#94423

X-original-commit: ba7a03b942c4d9991d8179a421925cea9a80ddf3
Signed-off-by: Max Whale (mawh) <mawh@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>