[FIX] data_cleaning: Fix Discard Button Issue Due to Many2one Field Refactor.
Currently, the discard button is not working correctly when trying to discard
a record in the deduplication feature.

Steps to replicate:
- Go to: Data Cleaning > Configuration > Deduplication.
- Open any Deduplication Rules > Deduplicate > select any record or directly
  click on the discard button.
  (Make sure at least one record is available in list view.)
- Click `discard` on top left.

Issue:
`It will load the page but not discard the record.'`

This is not work because of a recent refactor where the data of a `many2one`
will be an object rather than an array. (task-3547961)

The discard button in the deduplication feature was not working correctly
after a recent refactor in Odoo SaaS version 18.4. Previously (in SaaS 18.3),
the record.data.group_id many2one field was represented as an array. The
discard logic was written to access the group ID as record.data.group_id.

After the refactor, the record.data.group_id field is now represented as an
object. Thus, record.data.group_id became undefined, which made parseInt()
return NaN. Because of this, the application failed to identify the correct
group and records to discard, causing the discard action to silently fail and
only reload the page without discarding any records.

Summary:
This change ensures that the discard button correctly identifies the group
and record IDs to discard records.

closes odoo/enterprise#94082

X-original-commit: 48002e14288d16e0135bfdd8ffed98c021f15bec
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Lakhichand Borse (labo) <labo@odoo.com>