[IMP] appointment: support inverse capacity for user apt
Appointments support capacities for users since [1] in 18.4

As the total capacity is now shown and editable for these kinds of
events we want to support actually updating the capacity on them.
Allowing users to create an appointment "With Marc Demo, for 5 people"
from the back-end easily.

This can be done simply by re-creating the booking line with the correct
capacity, to remain in line with the way it works for resource
appointments.

This also introduces tests to ensure the default values in the gantt
view matches what's expected for the appointment and resource/user
selected, and that saving this value is indeed allowed.

As well as a more specific test on the inverse to make sure it does not
allow overbooking (which we cannot support), or specifying capacities
when capacities are not managed, for consistency.

[1]: bce7e94650c337a9046958a7689c81db5b2a4c73

task-4909129

Part-of: odoo/enterprise#92263
Related: odoo/odoo#222741
Related: odoo/upgrade#8262
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>