[FIX] mrp_workorder: backorder displayed quantity
**PROBLEM**
When creating a backorder, the quantity to produce during an operation is correctly displayed on the shop floor step. But when clicking to modify it, the pop over display the total quantity to produce, and not the quantity to produce in that specific operation.

**STEP TO REPRODUCE**
1. create a BoM of product with 3 or more operations
2. Create a Manufacturing order for i.e. 10 unit
3. Open shop floor
4. Register the production in shopfloor:
- Op1 – 10 units registered
- Op2 – 7 units registered
- Op3 – 5 units registered

5. At the end, a backorder is created for 5 units.
6. When we open the wizard to register the production on the Op2, the quantity to produce that is displayed is 5, which is wrong because we only need to produce 3 unit for that step.

**CAUSE**
In the setup() of MrpDisplayRecord, we wrongly set `quantityToProduce` equal to the quantity to produce for the whole MO. This doesn't work when the quantity to produce in each WO is different (which is the case for some backorders).

opw-5011739

closes odoo/enterprise#98775

X-original-commit: 68105110c289987140bc51657de40bc19df7e6d9
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>