[FIX] appointment: book only 1 seat per resource with inverse
[1] introduced the ability to use resources for more than 1 appointment
at a time even when not managing capcities.

As the computation is based on the capacity used by the resource, the
inverse method on total reserved capacity should only set 1 capacity
used per resource for a calendar event. Instead of the previous behavior
of considering "the whole capacity of the resource" is used.

[1]: bce7e94650c337a9046958a7689c81db5b2a4c73

task-4909129

Part-of: odoo/enterprise#92263
Related: odoo/odoo#222741
Related: odoo/upgrade#8262
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>