[FIX] pos_restaurant_appointment: notify only relevant table bookings
Task: [#5005216](https://www.odoo.com/odoo/my-tasks/5005216)

---
**Before:**
If no appointment type is specified in the POS config and a table is booked via the website,
the floor plan is notified of a new booking because the resource used is one of the POS config resources.
However, if the page is refreshed, the booking disappears since no appointment type is defined in the POS config

**After:**
The floor plan is notified of a new booking only if the appointment type of the booking matches the one specified in the POS config.
If no appointment type is set in the POS config, no booking notifications are sent.

Additionally, when the "Table Booking" field is unchecked in the POS config, the appointment type is automatically unset.

closes odoo/enterprise#99738

X-original-commit: 6d8212ec19aaf96ccb28b98c93251df87c547f34
Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>
Signed-off-by: Ludo Traina (ltra) <ltra@odoo.com>