[FIX] documents: fix search panel drag and drop
1. In the Documents app, drag any folder from the search panel
2. Drop it as subfolder to any valid destination folder
3. You get a traceback.

`user_folder_id` is a `Char` fields, so we made the backend
only support strings for it, including for
`actions_move_folder`'s `target` but failed to align the
frontend to this stricter API.

Applying the same rule for creating shortcuts for consistency.

Follow-up of 6bdcc357.

Task-5055163

Part-of: odoo/enterprise#93597
Related: odoo/upgrade#8335
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>