[FIX] stock_barcode: avoid packing new lines in barcode
Steps to reproduce:
- Go to the Barcode app and create a new transfer
- Add a Product, select any storable product
- Confirm

Issue:
The move line will be put inside a new package.

When creating the move line, it will also go through the
`_inverse_outermost_result_package_id`, as that field will be set with a
False value.
However, due to not checking if we set a value in that field, it will
try packing it with that value regardless, creating a new package.

Task-4314600

closes odoo/enterprise#93733

Related: odoo/odoo#225184
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>