[IMP] esg: automatically create assignation rule for emission factor in AML
Before this commit, when the user assigns an emission factor to an
account move line, no assignation rule is created based on that, which
means the user has to either manually create an assignation rule into
emission factor or each time select that emission factor for the same
product/partner set on other account move line.

This commit automatically generates an assignation rule into emission
factor when the emission factor is manually set into an account move
line. By doing that, the user will be able to use the assignation rule
to automatically assign that emission factor to other account move
lines.

task-4933207

closes odoo/enterprise#97736

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>