[FIX] ai_website_livechat: Fix ai_livechat website snippet bugs
1. move focus to component conditionally

The ai_livechat component has 2 chat styles namely fullscreen and popup.
The focus should be set on the textarea of the component only if the
chat style is fullscreen.

Steps to reproduce:
1. Set the chatstyle to popup on the ai_livechat component.
2. Enter a prompt.
3. A chatwindow pops up, and the AI response is shown in that window.
4. The focus is moved from the chatwindow to the textarea of the
   ai_livechat component.

2. The condition that checks the agent presence and the access rights to it
when creating a new ai chat channel for the ai livechat website snippet
was incorrect. It should use or instead of and.

3. The snippet options plugin (AILivechatOptionPlugin) of the ai_livechat
   website snippet should be added as part of the website.website_builder_assets
   instead of html_builder.assets because it is only valid in website context.

Part-of: odoo/enterprise#93686
Related: odoo/odoo#225089
Signed-off-by: Panagiotis Kyriakou (paky) <paky@odoo.com>