[IMP] hr_expense_stripe{,_demo}: add support for physical cards
Send a mail when the card is sent by stripe. The one receiving this
mail with be the person that ordered the card if it's delivered to
the company address or the person that will receive it.

We added a state 'pending' which needs to always be inactive on stripe
which allows to know the card wasn't activated yet and need activation.
This is to present users to easily activate the card while it's in
transit and put the responsability on Odoo.

task-4859836

closes odoo/enterprise#90988

Signed-off-by: Julien Alardot (jual) <jual@odoo.com>