[FIX] website_generator: always call clearInterval
In a component lifecycle, it may happen that onWillStart is called
but not onWillUnmount. Indeed, if the component is destroyed
before being mounted (because the current rendering has been
cancelled), onWillUnmount isn't called. As a consequence, in the
WebsiteGenerator component, the setInverval might never been
called, thus producing an orm call every 10 seconds, when the
component is destroyed. These calls lead to crashes ("Component is
destroyed").

A solution is to start the setInterval in onMounted.

closes odoo/enterprise#93771

X-original-commit: 9a3e9195c42ff2ac4a958bdf68f30e594d00a759
Signed-off-by: Géry Debongnie <ged@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>