[FIX] planning: correct ICS datetime conversion when no employee assigned
The test `test_planning_ics_file_without_assigned_employee` failed when running
without demo data because the slot timezone was `Europe/Brussels` while the employee
timezone was `UTC`.

The previous code in the method `ics_datetime()` converted to the slot timezone and
then relabeled it as the employee timezone with .replace(tzinfo=...), which shifted
the actual instant.

This change ensures that ICS datetimes are always converted using astimezone to a single target tz:
    - employee tz if the slot is assigned,
    - otherwise the current user tz or `UTC` as fallback.

The test was also updated to assert the correct fallback `UTC` values:
    DTSTART:20230602T080000Z
    DTEND:20230602T170000Z

runbot-231213

closes odoo/enterprise#94533

X-original-commit: 2e5672a96e096f1b09f38fe31e1d030a972ab9d8
Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>
Signed-off-by: Kawtar Drissi El Bouzaidi (kdeb) <kdeb@odoo.com>