[IMP] *_extract: improved boxes interface
This commit refactors and improves the OCR boxes interface used to fill
in the fields of bills and bank statements.
It is now available on all fields, instead of only the fields detected
by the OCR. The supported field types are: char, text, html, integer,
float, monetary, date, datetime and many2one.

It is now possible to select multiple boxes at once. The behaviour will
vary depending on the situation:
- On textual fields, it will concatenate the content of all boxes and
  set that value.
- For numeric and date fields:
  - on regular fields, it isn't supported as it doesn't make sense.
  - on x2many fields, it will create one line from each box.
- On top of this, if lines were created using the tool, it will search
  for aligned boxes among the selection to fill in the corresponding
  line.

The form view of bank statement lines was also modified to accomodate
these changes. The date field and the bank statement lines are now
displayed. The attachment was removed as it was redundant with the
chatter.

Because of the time constraint, the tests have been removed as they need
a rewrite, this will come in another commit later.

task-5055731

closes odoo/enterprise#93918

Related: odoo/odoo#225566
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>