[FIX] stock_barcode: convert uom in barcode package
Manual forward port of https://github.com/odoo/enterprise/pull/90878

**Problem:**
When scanning a package with a different UoM than the barcode line, the
conversion is not made.

**Steps to reproduce:**
- Enable the "Packages" setting;
- Create a new storable product and set kg as its UoM;
- In the inventory tab, add "g" in the packagings
- Click on the on hand smart button and select update quantity
- Add a new line;
- In the package column create a new package;
- Set a quantity of 10 kg;
- Create a delivery and select your product;
- Set a demand on 10000 and select g as the UoM;
- Mark as todo;
- Open the delivery in the Barcode app;
- Scan the package.

**Current behavior:**
The quantity on the line is now 10 / 10000 g

**Expected behavior:**
It should be 10000 / 10000 g

**Cause of the issue:**
https://github.com/odoo/enterprise/blob/4c9fa9dc010958710d848fbcb3241b17ea7205ca/stock_barcode/static/src/models/barcode_picking_model.js#L1500-L1505
remaining_qty is expressed in the uom of the quant so it will be 10 but
qty_needed is expressed in the uom of the line is it will be 10000.
qty_used beeing the minimum of those two it will be 10.

**Fix:**
To define how much quantity to take from the package, we convert the
line's quantity by using the package's UoM.
Then, when we add this quantity to the line's quantity, we re-convert it
by using the line's UoM.

opw-4860064

closes odoo/enterprise#95892

X-original-commit: 641be00de781530db09b2282d42c5d31ff086c6c
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Pierre-Louis Hance (plha) <plha@odoo.com>