[FIX] spreadsheet_edition: cannot follow non-store field
Before this commit, it was possible to follow a non-store field relation
in the pivot chain feature. It leads to an error in the cell as it's not
possible to retrieve the value of a non-stored field.

In order to ease the work of the user, we do not display non-store field
relation.

This commit also adds a missing test for the m2m relation.

closes odoo/enterprise#93701

Task: 5059203
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>