[IMP] account_reports: reset exceptions on tax closing
When we do a tax closing, we expect the tax lock date to be set so that
no entries can alter the tax closing. If however we applied an exception
on the tax lock date before doing a tax closing, that exception will
still exist afterwards.

This change makes sure that any exception on the tax lock date is
removed whenever a tax closing is done.

task-4525909

Part-of: odoo/enterprise#93327
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>