[IMP] accountant_knowledge: use `footer` option rather than `footer-right`
This commit updates the method for rendering the audit report PDF to use
the `footer` option instead of the `footer-right` option, which is being
removed in the COM PR.

To render the footer, we now use a custom template, as the template used
in the account report does not meet our needs. The account report footer
includes the company name, logo, and other elements we don't want. It
also loads numerous assets and applies a specific paper format. To avoid
complicating the existing template, we are creating a new template from
scratch.

Reference: odoo/odoo#222199

Task-4989809

closes odoo/enterprise#91906

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>