[FIX] hr_contract_salary_payroll: correct gross computation in offers
Reproduce:
In debug mode go to Payroll > Employees > Offers > Create a new offer
with a contract template that has no working hours. The gross salary
shown in the Salary Simulation Preview is 0.

Issue:
If the version has no resource calendar, its working hours are 0. This
sets the `work_time_rate` to 0, and when multiplied by the wage the gross
becomes 0 as well.

Fix:
Prioritize the version’s resource calendar when available, otherwise
fallback to the offer’s calendar, and finally to the company’s.

task-5051509

closes odoo/enterprise#95041

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>