[IMP] sale{,_stock}_renting: gantt planning improvements
This commit reworks the rental schedule/planning to make it editable.
Previously, the schedule was a static report view of the orders, which
made it impossible to reschedule, modify, or create new orders. This
limitation hindered logistics organization.

With this commit, we replace the custom report model with a Gantt view
of `sale.order.line`. The following improvements are introduced:

- Users can move order lines left or right to reschedule orders or
  adjust their duration.
- Users can move order lines up or down to change the product, customer,
  or reserved serial number.
- Users can create new orders directly from the schedule view.

task-4720184

See also:
- https://github.com/odoo/odoo/pull/224751

closes odoo/enterprise#88689

Related: odoo/upgrade#8347
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>
Co-authored-by: Victor Feyens <vfe@odoo.com>