[FIX] sign: prevent overlap between auto and drawn signatures in dialog
Fixed an issue where both the auto-generated and drawn signatures overlapped in the signature dialog.
This was caused by the system automatically setting the user's default signature to the drawn one when opening the dialog.
Now, the dialog no longer defaults to the drawn signature, preventing visual overlap and ensuring a cleaner user experience.

closes odoo/enterprise#93854

X-original-commit: ece2e5e2dc720cdc12b83b27f726824d12e4f114
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Moustafa Mohamed Ibrahim Tawfik Aly (ibmo) <ibmo@odoo.com>