[FIX] l10n_ec_edi_stock: prevent error when generating delivery guide without Barcode
Currently, an error occurs when generating a Delivery Guide if the Barcode
Scanner is disabled in the Inventory settings.

Steps to reproduce:

- Install the l10n_ec_edi_stock module and switch to the EC company.
- Uncheck Barcode Scanner in the Inventory settings.
- Create a new warehouse and set the Entity and Emission Point.
- Navigate to Inventory > Operations > Deliveries and create a new delivery.
- Add details > mark as Todo > Validate > Generate Delivery Guide.

Error:
AttributeError: 'stock.move.line' object has no attribute 'qty_done'

Root Cause:
At [1], the code references line.qty_done, but this field is defined in the
stock_barcode module at [2]. When the Barcode Scanner is disabled, the field is
not available, leading to the error.

Fix:
This commit updates the delivery guide values to use line.quantity instead
of line.qty_done at [1] and at [4]. Since in the stock_barcode module at [3],
qty_done is derived from quantity

[1]:
https://github.com/odoo/enterprise/blob/ba5b9790f28e2f7eabda22e5992737eab0e82c6e/l10n_ec_edi_stock/models/stock_picking.py#L354

[2]:
https://github.com/odoo/enterprise/blob/8c53e50df1cf9dc6d3ca4cae19c39135ac85d4e4/stock_barcode/models/stock_move_line.py#L23

[3]:
https://github.com/odoo/enterprise/blob/8c53e50df1cf9dc6d3ca4cae19c39135ac85d4e4/stock_barcode/models/stock_move_line.py#L48-L50

[4]:
https://github.com/odoo/enterprise/blob/8b72fdef63f634ccf436b49adbac5c1f9358c127/l10n_ec_edi_stock/views/report_delivery_guide.xml#L165

sentry-6851008674

closes odoo/enterprise#94826

X-original-commit: b752166548af16ddb13fc9af05471fef9672308f
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Utsav Maru (utma) <utma@odoo.com>