[IMP] stock_barcode: make partner instructions box scrollable
The partner-specific instructions div can take up too much space if it
contains multiple lines, so we move it out of the header into the stock
move list, so that we can scroll away from it with the rest of the stock
moves.

Plus cosmetic changes to the warning.

Task ID: 5060030

closes odoo/enterprise#93796

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>