[IMP] appointment: improve description
Purpose
=======
Improve event description by adding the booker details and
questions answers so that these info can easily be accessed
from the calendar popover and external calendars description.
Also displaying the description in the gantt and POS popovers
for quick access.

Technical
=========
Currently, at event creation, no description is saved into the event.
The get_attendee_description method is used to dynamically
generate the description only when it's needed to have up to date information.

However now that the description needs to be displayed
on every popovers, the event description field have to be
populated at creation. So bringing back the previous behavior
where the description was generated based on the form values
and directly set on the created event (revert of odoo/enterprise#47471).

A dynamic solution was investigated but couldn't be implemented:

- Populating the description field while still using the dynamic method
isn't a solution. Now that the description is displayed everywhere there
could be misunderstandings and data inconsistencies between what is shown
to the user on the front-end using the dynamic method (validation page +
mail templates) and what is show on the back-end using the description field.

- Changing the description field to a computed field is a bad idea as
the user custom notes entered manually in the description could be lost at
re-computation. There's no way to parse what has been manually added to what
needs to be re-generated.

- Tried to introduce a new computed field for the dynamic description part.
However external calendars API only have one dedicated html attribute meaning
we would have to merge the manual and dynamic description when syncing from
odoo -> external calendars but there would be no way to differentiate between
the 2 descriptions when syncing back from external calendars -> odoo.

Conclusion is that it's better to have a static description generated at creation.
If some infos were to be changed like the answers to questions or
partners details, the related info will need to be manually
updated on the description via the event form.

Task-4759201

Part-of: odoo/enterprise#89249
Related: odoo/odoo#217075
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>