[FIX] voip_ai: ensure call creation order in test
Before this commit
in `test_cron_transcribe_recent_voip_call_two_calls_at_same_time` test,
we were attempting to have two calls one created after another.
However since the creating of the second (earlier) call was based on the
absolute date, when the setUp took longer then anticipated the order was
the opposite of the desired.

With this commit we ensure that the 2nd call will have create_date before 1st
by pushing it back from the 1st call create date (not with creation of the 1st call)

closes odoo/enterprise#93807

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>