[IMP] iot: raise error downloading handlers unknown identifier
When an iot box tries to download drivers from the database, but the db
doesn't have a record corresponding to the IoT Box, the IoT Box receives an
empty string, making it hard to debug why it couldn't download handlers.

We now raise an unauthorized error to make it clearer.

closes odoo/enterprise#95934

X-original-commit: d372134941202b0b985d4c9fd3041e76c9efac1d
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>