[FIX] hr_payroll: fix UserError when canceling a payroll payslip
steps to reproduce:
-------------------
1. Install payroll
2. Create a user and grant "Administrator" access to Payroll.
3. Log in as the new user and try to cancel a 'Done' payslip.

issue:
------
A UserError is raised:
"Cannot cancel a payslip that is done."

observation:
------------
A user with Payroll "Administrator" access is unable to cancel a payroll payslip

cause of the issue:
-------------------
During cancellation, the system checks whether the user is "Admin"
instead of verifying if the user has Payroll "Administrator" access.
https://github.com/odoo/enterprise/blob/13832d80570956e504e1c09f41acbeb0bc4baedc/hr_payroll/models/hr_payslip.py#L509-L513

solution:
----------
Check that the user has Payroll "Administrator" access.

opw-5040029

closes odoo/enterprise#95277

X-original-commit: bd0851b75339e07cb19ea6f67b282cde341d5167
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>