[FIX] l10n_uy_edi: filter l10n_uy_edi.document from S&P wizard
When downloading attachment via the send & print wizard, we get an error
from the server. This is because we raise an assertion error if any
attachment is not from 'account.move' model. But the CFE file is from
'l10n_uy_edi.document' model.

With this commit, we extend the `_action_download` method to filter the
CFE file from the attachments. It is not blocking for client as he can
still download it from the form view of the CFE document.

Steps:

- Create an invoice
- Set a 0% tax on the invoice line
- In 'Other infos' tab, fill the 'Incoterm', 'Sales Modality' and
  'Transportation Rules' fields
- Confirm
- Open S&P wizard, select 'Create CFE' and confirm
- Reopen S&P wizard, select 'Download' and confirm (can be done along
  the previous step too)
-> Error

opw-5043902

closes odoo/enterprise#94458

X-original-commit: 7764df6b9950cf1f52c61b49304e9153a011203d
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>