[FIX] l10n_mx_edi: fix discount issue with global invoice
Steps to reproduce:
-------------------
* Create an invoice with the following line:
  * Unit price: 47.25, Qty: 1, Taxes: 16%, Discount 50%
* Confirm, create a global invoice
> Observation: Error
```
Code : CFDI40108
Message : El TipoDeComprobante es I,E o N, el importe registrado en el campo no es igual al redondeo de la suma de los importes de los conceptos registrados.
```

Why the fix:
------------
The issue occurs because the `descuento` value, originally 23.625, is now being
corrected to 23.615 which leads to `importe` having a value of 23.635 which
round up to 23.64 and not 23.63.
https://github.com/odoo/enterprise/blob/08564f3312c255f2f3ab95cef5a9bfc57727bd1f/l10n_mx_edi/models/l10n_mx_edi_document.py#L1111
https://github.com/odoo/enterprise/blob/08564f3312c255f2f3ab95cef5a9bfc57727bd1f/l10n_mx_edi/models/l10n_mx_edi_document.py#L1336

Before this commit https://github.com/odoo/enterprise/commit/39759babddc732a312ec5cd6a60a2f1819abc62c
the discount value was being rounded when corrected. It would end up being evaluated
to 23.62. To not bring back the issue the mentionned commit fixes we round the
discount when generating the global invoice cfdi values. Now `importe` will
have a value of 23.63 as `descuento` is rounded to 23.62.

We also remove the condition `if not global_invoice` as it ends up that this code
is also needed with global invoice. This code was also present in the previous
version and for the case of global invoices.

opw-5023597

closes odoo/enterprise#98511

X-original-commit: 459948332829e37bac5511f5e43c32a3e255d8c7
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Sarah Bellefroid (sbel) <sbel@odoo.com>