[FIX] l10n_uy_edi: ensure nom_item and description handle falsy values correctly
This commit makes a minor adjustment to the logic for extracting item names and descriptions in the _l10n_uy_edi_get_line_nom_and_desc method. The change ensures that the presence of aml.name (the line description) is checked directly, rather than relying on the truthiness of the entire aml object.
Before this, if the aml of an invoice only have the account_id field filled and no product_id, the _l10n_uy_edi_get_line_nom_and_desc will raise an error message like TypeError: 'bool' object is not subscriptable when trying to create de CFE document.
With this fix, we ensure there is a name to grab or the '-' will be send instead.

closes odoo/enterprise#94264

X-original-commit: 37f5627adced18876871447b7f0da57b6bfa83f2
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>