[FIX] ai: fix crash when chat closed before getting answer
If a user sends a message to an ai agent and then closes the chat
channel before receving the response, an error will be shown.

Cause of the error :
There is a series of errors that happen and only the last one is shown
to the user.
When the channel is deleted (closed), a serialization error occurrs
because one transaction is trying to delete the channel and another is
trying to post the ai generated response to the same channel.
The deletion operation is executed and the other transaction is retried.
the `/ai/generate_response` controller is called for retry. However,
it will throw a NotFound error given that the channel has been deleted.
This error is caught on the client side and then `/ai/post_error_message`
controller is called to post that error to the chat channel. This controller
will also throw a NotFound error and that error is shown to the user.

Note: No issue will happen if the response generation transaction is executed and the
deletion transaction is retried, because it will delete the channel
after the response was posted which is a normal behavior.

task-5063221

closes odoo/enterprise#93840

Signed-off-by: Abdelrahman Khaled (akha) <akha@odoo.com>