[FIX] knowledge: add activity filters in knowledge articles
Currently, when the user tries to open any activity of the knowledge
article, it opens all articles instead of the one which has an activity
assigned to them.

Steps to reproduce this issue:
1) Install the Knowledge module
2) Set up an activity for yourself on a Knowledge article
3) Open the activities from Activities (top left corner)

Issue:
You will end up in the all articles list, with no filters applied.

Cause:
When the user clicks on the activities, a default search
filter is added in the context, which is then applied on the view.

But in the knowledge module, we don't have any search filters
for the activities. Therefore, it renders all knowledge article records.

Solution:
Add search filters for the knowledge articles.

opw-4997201

closes odoo/enterprise#94411

X-original-commit: 5fcd6ad6fe02dc00df1b45c446889c9aad32f06b
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>
Signed-off-by: Altaf Shaik (alsh) <alsh@odoo.com>